Fix iostream.h include; incorrect variable names
`include <iostream.h>` is wrong, it's `include <iostream>` as used in the other parts.

Variable names starting with an underscore followed by an uppercase letter are reserved for the implementation, same for variables containing two consecutive underscores.
https://stackoverflow.com/questions/228783/what-are-the-rules-about-using-an-underscore-in-a-c-identifier

Note that the `using namespace std;` is actually frowned upon, but as it's so wildly in use (and also the default on ideone) I didn't change that.  Also there's some minor terminology mistakes but I doubt that the book benefits from fixing those.